+++ bin/run-examples.sh
@@ -6,8 +6,8 @@ cd "$(dirname "$0")/.." || exit 1
 
 
 uv run coverage run -a -m examples.simple -h
+uv run coverage run -a -m examples.simple 1 --baz 2
 uv run coverage run -a -m examples.simple foo 1 --baz 2
-uv run coverage run -a -m examples.simple_main 1 --baz 2
 uv run coverage run -a -m examples -h
 uv run coverage run -a -m examples foo 1 --foo2 foo --foo12 '{"foo": 1, "bar": "baz"}' --foo-bar 1
 uv run coverage run -a -m examples sub -h